configure: re-instate libsystemd-journal.so as fallback
For some older distros libsystemd.so is not available
(e.g. openSUSE 12.2), so the change commit 25d9bd6 means
we've gone from build the pcp-pmda-systemd rpm to not
building this rpm, which means the upgrade will fail.

This change maintains libsystemd.so as the preferred library
to check, but if that fails allows libsystemd-journal.so as
an alternative.