src/pmdas/sample/GNUmakefile: install executable and DSO PMDA binaries
The sample PMDA is designed to be built from source, but if the source
changes we and we re-install the package, then we may end up with
installed binaries that do not match the source ... and this can cause
QA failures.

Not sure why these were omitted from the install target, other than
possible to force checking that the PMDA can be rebuilt from the
package installed source, but that seems a stretch.