rc scripts: always use full paths during process lookup
The init scripts failing to start or restart daemons when other
applications have processes called pmcd, pmproxy, etc.

This is because the pcp.env function _get_pids_by_name() does
not check for the full path of the binaries and tries to kill
the foreign process.

In particular, this issue was observed with concurrent use of
UniSpool from http://aki-gmbh.com/en/products/unispool.html as
this has a pmcd process which breaks the start of PCP's pmcd.

This resolves Red Hat BZ #1365424.