debian/rules: fix for strange regression in some Debian builds
Recently, and only apparently on Ubuntu 16.04, the pmcd.stp file
was getting built with the correct mode, but the mode was changed
during packaging ... and qa/994 was catching that the installed mode
was not as expected (and fixing it) when it is run after any test
from qa/check.

This explains qa/000 failures on bozo.