qt tools: enable use of Qt5 toolchain, headers and libraries
Move forward with support for Qt5 for all Qt tools and libs
in PCP, and also latest Qwt 6.1.4 (also needed for Qt5).

Many API and ABI changes here.  A bit of fine-tuning remains
to reach some of the UI neatness from Qt4, but it's a clean
build for the first time at least, and the tools run.

(Qt5 required for PCP GUI tools under MacOSX El Capitan.)