src/pmlogger: fix logic errors in disconnect() handling
From within the callback handler, there was a disconnect() followed
by a fallthrough to code that assumed the preceding myFetch() had
worked ... on vm31 (x86_64 Fedora 23) this was triggering a SEGV
and qa/389 failure.

On inspection there were cases in which disconnect() was being
called multiple times even though no pmReconnectContext() had
succeeded (triggering unnecessary work and repeated and misleading
diagnostic messages), and a bit of looseness in the unlocking
of the c_lock context lock.