build: fix fedora.spec for Fedora el5 build
Fix a few snags I hit in the fedora builds for 3.11.5 el5 build.
On el5 there is no systemd and also no libvirt-python package in
the build root. These fixes are also committed in pcp.spec in the
el5 branch (only) of the fedpkgs repo, but don't seem to be needed
for Makepkgs builds using pcp.spec.in.