src/pmlogger: fix SIGINT issue
Issue raised as https://github.com/performancecopilot/pcp/issues/116.

This commit (and the earlier pduread() change in libpcp) fixes the
SIGINT part of the issue ... the root trigger is fetching lots of
metrics with a very short (10msec) update interval, which means
there is close to a 100% guarantee that the SIGINT will interrupt
a socket read() and the resulting error was not being propaged back
up the call stack correctly.

The mem leak part of the issue will be addressed separately.