pmwebd: correct HTTP time zone syntax in Date:/Last-Modified: headers
During testing of the checklist webapp prototype, it was found that
the HTTP-level caching of some of the documents was broken by a slight
error in the syntax of the RFC822-ish timestamps printed by pmwebapi
for file service.  We now use the hard-coded & correct "GMT" suffix
rather than rely on strftime("... %z").  Confirmed function with
web browser about:cache indicating proper cache life/expiry.