pmchart: fix signal blocking for the qwt picker events
Signal blocking was (implicitly) being done on the chart
object itself, instead of the picker object.  This meant
signals were not being blocked, such that every chart is
active at some point during a selection event.

This was benign before, but soon we wish to popup a new
metric detail dialog box and we need to know which chart
is actually active (and just the one selected thanks!).