src/libpcp/src/pmns.c: change logic in __pmHasPMNSFileChanged()
Changes elsewhere have meant that pmcd reconfiguration step are
now significantly quicker (this is good), but this has exposed a latent
design error in __pmHasPMNSFileChanged().

The check for "has the pmns changed" is based on stat()ing the root
PMNS file and checking the mtime ... unfortunately this has only second
resolution on some systems, which was OK when PMDA Install/Remove was
slow ... but not now.  Improved the logic to check the file size and
the mtime.

Also needed another static inside the lock guard, so update check-statics
accordingly.