src/pmcd/pmdaproc.sh: tweak signal sending logic
Nathan's last change made delivery of SIGHUP more robust.

Unfortunately it widened the window between pmcd getting the SIGHUP
and pmcd doing the post-SIGHUP work (this is delayed in pmcd so it does
not run in a signal handler, i.e. interrupt, context).  This increased
the chance that the PMDA install script would fail to find the recently
installed PMDA's metrics because pmcd had not yet finished its work.

After the loop checking for the delivery of the SIGHUP, re-instate
a pmsleep $signal_delay to compensate, and this has the side-effect
that it can be over-ridden from the PMDA's Install script if needs be
(which reverts to the former behaviour).