src/libpcp/src/auxserver.c: be a little more tolerant of EADDRINUSE
With recent changes to the daemon restarting logic to speed things up
we seem to have opened a window where we try to bind to a specific
port (e.g. pmcd's client connection port) before the last use has
been fully torn down, resulting in EADDRINUSE errors on restart.

The problem seems particularly bad on my QA vm19 running x86_64
openSUSE 12.2 ... but that may have more to do with this VM than
this distro.

The change here is to add a small delay (of between 1 and 4 x 250msec)
when bind() returns EADDRINUSE ... this seems to be enough to let the
previous socket usage be completely cleaned up and the new daemon is
off to the races.  Most calls see no delay.  Non-transient problems
now take 4 failed attempts and 1 second longer to report EADDRINUSE.