src/libpcp/src/context.c: tweak to __pmCloseChannelbyFd() and __pmCloseChannelbyContext()
These routines get called from multiple places because __pmGetPDU()
is used in lots of places, but it really only makes sense to clobber
to the underlying socket in cases where the caller is a PMAPI client.

So the "close socket" part is NOT done if __pmGetInternalState()
returns PM_STATE_PMCS, which means any socket between pmcd (or dbpmda)
and a PMDA is not closed if a socket read timeout leads us to the
__pmCloseChannelby*() routines.  In these cases, the higher level
error handling is sufficient.

Of course, if Frank's un-multiplexing of the client-pmcd socket on
the client side is fully implemented, all of this code is likely to
be culled because the motivation will have been removed.