qa/895: remove hotproc metrics as well and filter out pduread diagnostic
Just to be sure, to be sure, the hotproc metrics are dropped like the
proc metrics ... these are not needed for the integrity of this test.

While the test is trying to force a SIGINT to pmlogger while it is
waiting for a PDU response from pmcd, we cannot guarantee this will
happen.  On a small percentage of test platforms the SIGINT sometimes
arrives when pmlogger is not calling pduread(), and so the diagnostic
is not generated ... simplest to filter the diagnostic from the .out
file ...  it remains in the .full file if required for triage.