qa: be firmer about stopping/restoring the systemd Restart=always behaviour
Many QA tests need to take control of the stopping and starting of
pmcd (and pmlogger and pmmgr), but the recently introduced systemd
Restart=always behaviour is getting in the way.

So for the tests that care, force systemd to not restart one or more
of the pmcd daemons, and restore the systemd configuration at the end
of the test.