pmdaxfs: use persistent dm and md device names for xfs.perdev
dm-xx device names are not persistent. So use the mapped name for
these devices as the instance names in xfs.perdev.*. Also now uses
the full path to the device, since that could be ambiguous otherwise.
This is now also consistent with the device instance naming used for
filesys.* metrics.