qa: refinements to recent pmdalinux memory metrics
Checking the kernel code, found the buddyinfo metrics are
exported as unsigned long integers so update the metadata
to reflect that.  The QA test needed some tweaks too as it
was evidently run as root when generating the new .out but
that means certain access-control testing was skipped, and
the output was non-deterministic when run unprivileged.