src/libpcp/doc: data files for mk.cgraph
These files list the routines external to libpcp that are known
to be thread-safe or potentially thread-unsafe.

The posix.unsafe list is stable and from the POSIX documentation.

The glibc.{safe,unsafe} list is being built as mk.cgraph encounters
routines that have not been seen before, so this is evolving and only
includes glibc routines we use in libpcp.