Utilizations from rates are between 0 and 1 not 0-100 percent
The thresholds for these metrics assumed a percentage value.  However,
the rate function used in the derived metrics is utilization between 0
and 1 rather than 0 to 100 percent.  Thus, the test would always
return false.  Rescaled the threshold to the proper range.  This is
exposes one of the weaknesses of this encoding.  A static range check
would have flagged something like this to be always false, but no
sanity check like that is done.