src/libpcp/src: add __pmLock_extcall
First round of changes for this new mutex.  It is designed to be the
last on the call hierarchy (no other locks can be acquired while
__pmLock_extcall is held) and is designed for short-duration locking
while external routines that are not thread-safe are called.

This reduces some of the use of __pmLock_libpcp, so we're able to
focus on cleaning up the use of that mutex later.