pmdalio: ensure pmda sees config changes, and correct LUN size calculation
By default the tcm module holds a cache dict of storage object so it only
walks configfs once for device discovery. This is fine, if adds/removes are
done through the same system - but in the case of the pmda the cache is
a problem since it won't reflect changes to configfs correctly. This fix drops
the cache (bs_cache) at each refresh interval.

In addition a minor change was made to the calculation of LUN size to round
up (4.8GB = 5GB)