src/pmns/Rebuild: rebuild pmns if root is empty
If $PCP_VAR_DIR/pmns/root exists, but is empty then Rebuild would fail.
Extend the existing check when creating a new skeletal root to include
the cases where the root file is missing or the root file is empty.

Found in the QA farm, where an unscheduled shutdown after a PCP
install had left the root file empty at the next reboot.