qa/370: changes for sysstat v11.2.0
New data file and script changes to accommodate.

Also recreating the data file on a machine that supports active VMs
exposed a problem with the "user" time reported by sar -u ...  this
would appear to be usr + guest (thanks Nathan), so use derived metrics
to reconcile the values (like already used in the sys CPU case).