RHBZ 1403095, RHBZ 1411720: Archive interp mode performance:
Replace the linked list of instances for each pmid with a hash table.

This results in an additional improvement of 50% for the test case
for RHBZ 1403095 on top of earlier improvements. Running time is now 25%
of the original.

This results in an additional improvement of 80% for the test case
for RHBZ 1411720 on top of earlier improvements. Running time is now
7% of the original.

Also fixed an error in __pmFreeInterpData(). The hash table buffer for
the pmids was being freed while it was still being used.