Update pmdaproc ttyname to iterate through /dev
Iterating through /proc/$PID/fd's to look for various tty's was
causing a variety of selinux denials.  To the extent that it was
inappropriate to include that number of dontaudit lines in the type
enforcement file.

Instead, we now iterate through /dev/ and check if a device matches
the particular pid of interest.  Similar to the approach ps takes.