Better systemd integration, part 1
Changes to the pcp "init" script (which is not a systemd service)
to control the pmcd and pmlogger services with systemctl if that
seems appropriate, else revert to running the classical pmcd
and pmlogger init scripts directly.

Similar change to pmdaproc.sh for controlling pmcd during PMDA
installation.

This is part 1, because there are hundreds (I jest not) of QA tests
that are changed to provide better integration with systemd for
platforms where systemd is in use.  The QA changes will be in part 2
that will be committed in a day or so after some additional testing.