containers: add a pcp-standalone app container layer
Implements a second docker layer for PCP containers, such
that applications running in the cloud with no privileges
can have a minimal unprivileged pmcd co-process exporting
metrics from within the container if they so desire.

Sample containers for the Apache and Nginx web servers are
included here, which are also configured to run the PMDAs
for those services (with stats auto-enabled) too.