qa/964: refactor
This test has fallen on hard times in the wake of the pmlogger changes
to allow it to survive pmcd restarts, and somewhat dodgey logic that
was associated with counting pmlogger instances.

Reworked to run two PMDA installs ... one does not require a pmcd
restart (xfs) and one does require a pmcd restart (dm).  Check the
same primary pmlogger continues to run through both PMDA installs.

Also needed some extra care in the notrun guards and the trap
handling of the PMDA cleanup.