containers: an instrumented Java application standalone container
Adds the sample pcp-java-app container to the build, which shows
how to run a Java service in a container with pmcd, pmdammv, and
Parfait instrumentation all in place alongside it.

Several other aspects of pcp-base and other standalone containers
are tidied up here.  Using the Dockerfile COPY directive seems to
result in much simpler, clearer Dockerfiles - so more use is made
of it now.

Improved notes on using docker inspect to extract the docker run
command (see comments at head of each Dockerfile) avoiding having
to specify it in multiple places.  Switched to using ENV to avoid
any direct editing of installed configuration files, for PCP_USER
and PCP_GROUP.