src/pcp/pmcd/rc-proc.sh: add mkdir_and_chown()
This helper function provides the moral equivalent of mkdir -p,
except that for any directories created there is an option to also
chown each of these.

Intended for the {pmlogger,pmie}_{check,daily} scripts where we
are doing
	mkdir -p $path
	chown $PCP_USER:$PCP_GROUP $path
which is not quite right of $path contains multiple directory
levels that do not currently exist.

This starts to address one of the problems reported by Martins Innus
<minnus@buffalo.edu> on 31 Jan 2017.