qa/964 et al: finally fix random failures
Problem was the choice of PMDAs! Test previously used xfs and dm
but if the primary logger was logging metrics from one of these
PMDAs and that PMDA was Removed (as qa/964 does), then at some
time later pmlogger notices pmcd has restarted, reconnects, checks
the metrics finds the missing ones and exits ... this behaviour is
non-deterministic and depends on local primary pmlogger configuration
which is why the test was unreliable.

Tweak the qa test_perl PMDA to conditionally force a pmcd restart,
and then use this for both test cases.