pmdabind2: comment out bind2.zones.serial for now
The instance domain evaluation is non-deterministic, and
test qa/1087 is showing incorrect values being associated
with the different instances at times.

Looking at the different bind XML formats seen so far, we
will need to rework this a fair bit.  Since this doesn't
seem to be a critical metric, and Lukas is busy working on
rewrite of some aspects of the PMDA affecting this for the
next release, I've commented out this metric for now.