First cut at getting an osxSierra base image working with Vagrant.
This grabs an OSX base box from the interwebs and applies a bunch
of extra brew installs to get it closer to a compilable state.

Still doesn't work, but it's at least mimicing where I got to on my actual
home laptop.

Once we get closer, this could be used as the basis of a newer base box
snapshot so the amount of install time (the extra brew installs takes 20
minutes or more to complete).