Update the memory check to use mem.util.available
The mem.util.available is a more appropriate measure that mem.freemem.
Linux tries to make use the memory to cache files and and buffer data
being written to block devices.  As a result, the mem.freemem may lead
one to beleive there is memory pressure when it doesn't exist.