libpcp: new implementation for derived metric parser
The old parser used a homebrew lexer and FSA parser ... the RFEs for
changes in the derived metrics expression language had outstripped
the implementation.

This commit tosses all the old lexer and parser, and wheels in shiny
new ones based on bison/flex (really yacc/lex!).

With this commit there are no regressions in the QA coverage (after
the companion QA commit), but some of the newer language features
(like unary negation and boolean negation) are in the parser, but
not yet plumbed into the semantic analyzer nor the execution engine
... pro tem using these will return PM_ERR_NYI.