derived metrics - add check ? foo : bar ternary operator
libpcp changes, man page changes and QA in this commit.

Just at the moment, the semantic restrictions make this not quite
useful for operand expressions that are constant (there is no way,
just yet) to set the metadata for a constant, so you're likely
to see type, or units/scale semantic errors.