libpcp: ensure correct ctx number passed for local contexts
Resolve issue on code paths where multiple local contexts are
created, which resulted in no connection attributes being set
for contexts after the first.  This in turn results in a lack
of permissions on e.g. duplicated contexts which results in a
failure of torture_indom when run as root with Linux slabinfo
metrics (which require root access, so didn't failure before).

Update test qa/232 to run the slabinfo checks as root which is
a good reproducible test case for this bug.

Resolves Red Hat BZ #1430248.