pmmgr workaround for github issue/195
Python pcp clients generally do not tolerate an "-h HOSTSPEC" suffix
argument of the form that pmmgr likes to create for monitor daemons.
(This is because of shortcomings in the pmapi/python argument parser
bindings.)

For these programs, pmmgr now understands a "monitor-host-env"
configuration alternative.  Here, the $PCP_HOST environment varable
is used instead of "-h HOSTSPEC" to identify the target host.  Even
though the behaviour is undocumented, many C PMAPI programs also
support this (via the pmGetOptions family).