selinux: RHBZ1433271 adjust for other out of tree policy packages
Some of the contexts we adjust for, can be present in the fedora
buildroot, however not in the default policy.  Like pcp-selinux,
shipped after the fact, with their respective rpms.  We were erroring
out when trying to install the pcpupstream.pp, and those rpms (and
thus their policy packages) were not installed.

In this case, still check if the contexts are available, and set in
gnulocaldefs, however, add trigger{in,un}'s for the packages they rely
on, and split those type enforcement rules into their own, smaller,
policy packages.