perfevent_pmda: Fixes a sigsegv for test_dynamic_events_config
The perfevent pmda code initializes/parses a cpumask for each
PMU. Previously, it used to allocate a memory block with max available
cpus from archinfo->cpulist. And, then, it finds out the list of
allowable cpus on which events of a certain PMU can be monitored. That
causes an issue with the test case test_dynamic_events_config, in which,
the archinfo->cpulist initialized (from the system) may be less than the
cpumask available for a pmu in this test case. If that's the case, it
will cause a sigsegv, since, it tries to parse more cpus than the
allocated space.

This patch fixes this issue by allocating the appropriate space by first
looking at the cpumask and then parsing it.

Signed-off-by: Hemant Kumar <hemant@linux.vnet.ibm.com>