impl.h: add back link in __pmContext for handle, add lock assertion macros
Adding c_handle to the __pmContext struct avoids a whole raft of
lock inversion issues associated with calling __pmPtrToHandle()
once the c_lock is held (__pmPtrToHandle() does not need any locks).

This does break the ABI, but I'm not sure there is anything that is
built outside of the PCP packaging that would be exposed ... if this
is not the case, we need to roll the libpcp version.

The lock assertion macros are no-ops (other than code annotation)
unless libpcp is compiled with -DBUILD_WITH_LOCK_ASSERTS