libpcp: rework for "current context"
Previously we maintained the current context as thread-private data
holding the (above the API) context number (or handle).

This causes ugly lock inversion problems when we try to get a pointer
to the current context's __pmContext within libpcp.

The fix is pretty simple, the current context is now maintained as
a mirrored pair (curr_handle and curr_ctxp) both as thread-private
data in context.c.  And a new routine __pmCurrentContext()s that is
lock-free and exposes curr_ctxp for callers in libpcp.

__pmCurrentContext() is equivalent to __pmHandleToPtr(pmWhichContext())
except it requires no locking and does not acquire the ctxp->c_lock
as a side-effect.  The places where __pmCurrentContext() is used
are ones where the relevant context is typically already locked,
sometimes a long way back up the call stack.

All this works because although contexts[] may get relocated,
contexts[i] does not change once the context is established because
the __pmContext struct is never moved or free'd, and so curr_ctxp
has long-term validity, even if contexts[] has been relocated.