libpcp: more threadsafe changes ... pc_lock rework and unnesting c_lock
I think this commit addresses the smoking gun I've found that explains
the bad press that has been around for some time concerning the
multiplexing of a single channel from multiple contexts to the same
pmcd instance.

In the __pmPMCDCtl struct we have pc_lock to ensure serialization
of all the IPC over the channel (pc_fd) from a specific context to a
particular pmcd and back for a complete PDU round-trip.  But the other
fields of __pmPMCDCtl were being accessed without concurrency control
... most importantly (but not exclusively) pc_refcount ...  leading to
badness in the presence of concurrent calls to pmDestroyContext()
and pmNewContext() for the contexts associated with the same pmcd.

These changes leverage pc_lock to also provide concurrency control
over all the fields of __pmPMCDCtl.

There are other changes here to refactor some routines (especially
pmGetArchiveEnd() and friends) to remove nested locking of the c_lock
mutex.  There is more to be done here, the mutex is still recursive
at the moment.