libpcp: major re-engineering of internal context management
Previously there was a identify mapping between a context handle (say,
x) above the PMAPI and the associated internal __pmContext managed
within the library and accessed via contexts[x] within context.c
(this relationship, by design, is hidden from the rest of libpcp via
the functional abstraction provided by context.c).

The problems with this implementation are many, and include:
- ugly and hard to fix races between pmNewContext/pmDupContext and
  pmDestroyContext (they may be fighting over the same __pmContext
  in ways that are very difficult to catch)
- unnatural acts in libpcp to try and catch a context that is being
  used by one thread, but is the target of pmDestroyContext in another
  thread ... one could argue this is dodgey client design, but that's
  a weak argument if it happens by accident and libpcp borks
- the (impossible to detect) case where one thread gets a handle
  (say, x), someone (the same or a different thread) destroys that
  context, and someone creates a new context ... this will likely be
  x and the poor first thread maybe left clueless about the change

So, I've changed the implementation.

New context handles are now monotonic increasing for the life of a
process ... immediately solves the last problem above because the
new context is guaranteed to be >x and poor thread's handle x will
return PM_ERR_NOCONTEXT as soon as it is used as the current context
for a PMAPI call.

This does change the behaviour of the PMAPI, namely before given
the sequence:

    x = pmWhichContext()
    pmDestroyContext(x)
    y = pmNewContext(...) or pmDupContext()

it was highly likely (but not guaranteed) that x == y.  Now that will
never be the case.  Fortunately I can find no documentation that
suggests how PMAPI context handles are generated, and it's hard to
image anything but the most broken and brave PMAPI client assuming
the old behaviour.

The memory foot print of allocated __pmContext structures remains the
same (a high water mark defined by the maximum number of concurrently
active contexts).

There is one level of indirection (via contexts_map[] that shares
the same concurrency control mechanisms as contexts[]) and this is
used to map from a context handle to the (possibly much smaller)
array contexts[] to find the associated __pmContext.

This commit just makes the contexts_map[] change.  There will be
follow-on changes that will simplify the implementation of context
creation/destruction considerably.