qa: remade .out files and qa/200 filter tweak for context handle change
Context handles are now monotonic increasing ... a small number of
tests reported the context number in the presence of pmDestroyContext()
(without destroying a context the old and new behaviours are the same).

And some tests use -Dcontext where the format of the diagnostics
from __pmDumpContext have changed a little (to report contexts_map[]
indirection).