libpcp: refactor locking in logutil.c
Replace big __pmLock_libpcp mutex by local logutil_lock mutex to
protect the global pc_hc hash table.

Also dropped the libpcp lock around the readdir loop (dirp is on
the stack, so this is threadsafe).  And replaced __pmLock_libpcp
with __pmLock_extcall around some external calls to libc routines
basename() and dirname() that are not threadsafe.

And corrected a locking oversight introduced when I
changed from __pmHandleToPtr() to __pmCurrentContext() in
__pmLogCheckForNextArchive(), namely the need to remove the
PM_UNLOCK(ctxp->c_lock) call on the code path that was found by
helgrind and qa/4751.