src/libpcp/context.c: remove PM_CONTEXT_FREE
The sentinal was in the wrong place (using the c_type field in the
__pmContext) which could only safely be accessed when holding the
c_lock mutex and if we were to do that it would mean acquiring that
mutex in places that would introduce lock inversion and potential
deadlock.

The recently introduced contexts_map[] array provides a better place
to hold this state, and this is guarded by the contexts_lock mutex
which already held in the places where we're scanning contexts[]
and need to make sure the associated __pmContext is or is not free.

With this change the "race or deadlock" between pmNewContext() (or
pmDupContext()) and pmDestryContext() can be completely removed by
using the new MAP_FREE and MAP_TEARDOWN sentinals for contexts_map[].

Fixing this problem (and the monotonic increasing handle numbers
introduced earlier) means we never have to destroy/recreate the
c_lock mutex in the __pmContext ... and this removes another ugly
raft of data race and lock instability issues