libpcp_pmda: rework and document string handling for pmdaDSO and pmdaDaemon
Recent gcc changes have exposed a long-standing bug; it was observed
in the linux PMDA, but is really a day zero issue in the underlying
libpcp_pmda initialization routines.

Specifically we never documented how the helptext argument to pmdaDSO
and pmdaDaemon, and the logfile argument to pmdaDaemon are used.  Some
PMDAs assumed the storage needs to persist after the pmdaDSO/pmdaDaemon
call and some did not.  The former always worked, the latter (e.g. the
linux PMDA) worked by luck until their luck ran out recently.

This commit adds strdup() calls in the libpcp_pmda routines and
documents in the man pages that the strings are copied and so may be
subsequently released by the caller.