libpcp: add missing mutex unlock in pmLoadDerivedConfig()
Spotted by Nathan (thanks) during a code merge review, this has been
there forever.  It is (mostly) benign because the mutex in question is
(today, not at at some point in the future) recursive, and almost no
one (much less a threaded someone) uses pmLoadDerivedConfig().

Also added a new QA test (954) to demonstrate the bug and confirm
the fix.