src/libpcp/config.c: remove nasty recursion
Previously we had this possible call cycle
    pmGet*Config -> pmgetconfig -> __pmconfig (fatal error) ->
    pmprintf -> pmGetOptionConfig

this was known, and there was an attempt to detect the (infinite!)
recursion in __pmconfig() ... but that code was badly borked in a
multithreaded environment, which led to decompression failing over
in fopen_securetmp() when PCP_TMP_DIR could be looked up and the
temporary file creation failed and we return a bad fd and note this
(but the diagnostic is hidden behind a (pmDebug & DBG_TRACE_LOG) guard)
and fopen_compress() returns NULL and then either _logpeek() returns
NULL or __pmLogChangeVol() returns -oserror(), which we previously
set to -2 (No such file or directory) way back in fopen_compress().

AND this is the root cause of the mystical
    ./archives/20130706.meta -2
not
    ./archives/20130706.meta NUMBER
in qa/4751 and qa/953 and ... guess what, archives/20130706 is the
only bzip2 compressed archive that is used by these qa tests!

Rather than trying to fix the recursion detection (that I don't know
how to do in a non-intrusive, thread-safe, manner), I've just replaced
the "this is really fatal" message output from pmprintf() to fprintf()
... no recursion and no problem any more.