src/libpcp_pmda/callback.c: fix small regression
When the metric search changes were made, there was a small regression
on the error path for the (very rare) case where a metric is in the
PMNS long enough to reveal a PMID, but using that PMID in a fetch
fails because the PMDA has amnesia.  Might happen for a real PMDA with
dynamic metrics that come and go.

Does happen for the broken PMDA in QA, and this has been causing qa/274
to fail for some time.