build: create dso variant of pmdasimple as well
One of our GSoC students noticed that we do not create the
pmda_simple.so file in a git repo build, which is slightly
confusing when reading the dbpmda(1) EXAMPLE section while
using a git repo build.  The installed pmda_simple is fine
and always has been, just the git build makefile.