assorted: rework for change in context handle semantics
Synchronization of the fetch profile now uses the [internal]
context slot index (index into contexts[] on the client side which is
relatively small) in place of the [above PMAPI] context handle (which
is monotonic increasing).  The protocol for keeping the client-side
and pmcd-side in sync is unchanged, but there is no memory leak on
the pmcd side.

Preserve Dave's hashing on the pmcd side  ... with one slight logic
change to accommodate context slot index numbers being reused
by a single client, e.g. by pmNewContext ... pmDestroyContext
... pmNewContext (or pmDupContext).  This ensures the number of
profiles held on the pmcd-side is no more than the maximum number
each client has created for this pmcd, which offers an improvement
memory footprint over the original linear array scheme.