pmdadm: add device-mapper stats metrics - second patch
This is the second patch from Fumiya Shigemitsu <gyro.scaremonger@gmail.com>.
This makes the dmstats code conditionally compiled on HAVE_DEVMAPPER and
HAVE_DMSTATS. Note some older versions of lvm2 do not have the dmstats
interface, so the dmstats code is conditional on HAVE_DMSTATS. The dmstats
support was added in version 2.02.166 about a year ago, so should be fine
with Fedora >= 25 or Ubuntu >= Zesty but earlier releases may not have support.

The HAVE_DMSTATS macro is not in the configure script yet (coming in the next
patch).

Other changes are :
- include libdevmapper.h (from device-mapper-devel) and link with the library
  $(LIB_FOR_DEVMAPPER), as defined in src/include/buildefs.in

- remove the duplicate declarations of dm_stats_metric_t and dm_stats_counter_t
  since these are already defined in libdevmapper.h when we HAVE_DMSTATS

- remove the _count sufix from some of the metric names (since PCP has metadata
  that already designates the semantics of each metric as either instantaneous
  or counter and so the _count suffix isn't needed)