pmprobe: add -F option for further pmlogconf speedups
Removes another series of pmcd roundtrips in pmlogconf use
of pmprobe, this time by removing unneeded single namespace
entry lookups within pmprobe (defer to pmLookupNames now in
this case).  This is good for removing multiple seconds of
elapsed time on remote pmlogconf hosts.

The next optimisation point is a pmLookupDescs API (variant
that accepts multiple PMIDs), which is also going to yield
shavings in the multiple-seconds range for remote hosts and
multiple metrics at once (the pmlogconf default mode now).