libpcp: major refactor to unwind nesting of c_lock locking
We need to push the current __pmContext * deeper down the callstack
and then allow internal routines to only take the c_lock if it
is not already locked.

For example: pmGetInDomArchive

This routine is callable from above the PMAPI (so without holding
the c_lock), but is also called internally from __pmLogFetchInterp()
with the c_lock mutex held.

Refactoring using a new pmGetInDomArchive_ctx() with an additional
first parameter that is a pointer to a __pmContext ... if this is
NULL we need the "not locked" mode, otherwise we need the "already
locked" mode. So the real work is done in pmGetInDomArchive_ctx()
and pmGetInDomArchive() becomes a wrapper like:

int
pmGetInDomArchive(...)
{
    ...
    sts = pmGetInDomArchive_ctx(NULL, ...)
    ...
    return sts;
}

Other PMAPI routines similarly "wrapped" in this commit:
    pmNameAll()
    pmLookupName()
    pmLookupDesc()
    pmNameInDom()
    pmFetch()
    pmUnpackEventRecords()
    __pmSendResult()
    __pmDecodeResult()
    __pmDumpResult()
    __pmDumpHighResResult()
    pmGetPMNSLocation()

Plus some internal static routines (where we don't need a wrapper to
preserve the old API):
    eval_expr() in derive_fetch.c
    bind_expr() in derive_parser.y.in
    GetLocation() in pmns.c
    dump_valueset() in util.c

I expect this pattern may be needed elsewhere are we move towards
making the c_lock mutex no longer recursive.