golang: bind pmUseContext and context operation synchronously
Introduces locking around calling an operation that operates on a
context (like pmFetch) and using the context with pmUseContext.

This doesn't appear to have a performance impact for a single context
and a single thread, but it's possible it may be unacceptable for
multiple threads/goroutines. It is possible that we could introduce
a different constructor or an option if the user knows they will only
ever have a single PmapiContext and its safe to bind these two together.