qa/347: take better control of primary pmlogger
This test was exposed to a change in behaviour that stems from allowing
pmlogger to try and reconnect when pmcd goes away ...  if the pmcd
is reconfigured then there is a possibility that the pmlogger will
persist and when pmcd is restarted the pmlogger notices the metric
metadata or availability has changed and pmlogger exits.  The exit may
race with a pmcd+pmlogger restart and a QA test (or a check callback)
that checks for the primary pmlogger to be running again may fail.

The fix is to explicity stop the primary pmlogger at the start of the
test and make sure it is restarted in the trap handler.  This is OK
as this test is _not_ designed to check pmlogger resilience.

There may be other tests that have a similar exposure.