qa: take better control of removing $seq.full
Rather than doing this in each script, check (and hence recheck) now
do this before running each test.

Even if a test is not using $seq.full, there may be a check callback
script in play that could be appending to $seq.full ... without this
change some of the $seq.full files were growing as individual tests
were repeatedly run.